Revert "架构补充"
