+++ src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.application.name=acnhdb
+spring.datasource.url=jdbc:mysql://localhost:53306/acnh
+spring.datasource.username=test
+spring.datasource.password=StrongPassword123!
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
