--- src/main/resources/application.yaml
@@ -1,19 +1,0 @@
-spring:
-  application:
-    name: acnhdb
-  datasource:
-    url: jdbc:mysql://localhost:53306/acnh
-    username: test
-    password: StrongPassword123!
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL8Dialect
-
-
-mybatis-plus:
-  mapper-locations: classpath:/mapper/**/*.xml
-  type-aliases-package: com/ecnudbcourse/acnhdb/entity
